% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{createSampleID}
\alias{createSampleID}
\title{Generate the samples ID for the simulated dataset.}
\usage{
createSampleID(nbGeneration, nbSample)
}
\arguments{
\item{nbGeneration}{a positive \code{integer}, the number of generations
simulated.}

\item{nbSample}{a positive \code{integer}, 
the number of controls (CTRL) and cases in the simulated dataset.}
}
\description{
Generate the samples ID for the simulated dataset. The
standard format of the samples ID is : 
"F[Number for the generation]_[Number for the sample]
_[OC for case or C for control]"
}
\examples{

## Create sample ID 
methylInheritanceSim:::createSampleID(nbGeneration = 3, nbSample = 6)

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
