% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSim.R
\docType{data}
\name{dataSimExample}
\alias{dataSimExample}
\title{TODO}
\format{a \code{list} containing:
\itemize{
\item \code{stateInfo} a \code{GRanges}, a synthetic chromosome as 
generated by \code{getSyntheticChr} function. 
\item \code{stateDiff} a \code{list} containing:
\itemize{
\item \code{stateDiff} a \code{vector} of \code{integer} (\code{0} 
and \code{1}) with length corresponding the length of \code{stateInfo}.
The \code{vector}
indicates, using \code{1}, the positions where the CpG sites are
differentially methylated.
\item \code{stateInherite} a \code{vector} of \code{integer} (\code{0} and 
\code{1})
with length corresponding the length of \code{stateInfo}. The 
\code{vector}
indicates, using \code{1}, the positions where the CpG values are
inherited.
}
\item \code{treatment} a \code{vector} of \code{integer} TODO
\item \code{sample.id} a \code{list} of 3 \code{list}. Each of the 3 
\code{list} contains a \code{list} of 12 entries each containing a string
of \code{character}, the name of the sample. TODO
}}
\usage{
data(dataSimExample)
}
\value{
a \code{list} containing:
\itemize{
\item \code{stateInfo} a \code{GRanges}, a synthetic chromosome as 
generated by \code{getSyntheticChr} function. 
\item \code{stateDiff} a \code{list} containing:
\itemize{
\item \code{stateDiff} a \code{vector} of \code{integer} (\code{0} 
and \code{1}) with length corresponding the length of \code{stateInfo}.
The \code{vector}
indicates, using \code{1}, the positions where the CpG sites are
differentially methylated.
\item \code{stateInherite} a \code{vector} of \code{integer} (\code{0} and 
\code{1})
with length corresponding the length of \code{stateInfo}. The 
\code{vector}
indicates, using \code{1}, the positions where the CpG values are
inherited.
}
\item \code{treatment} a \code{vector} of \code{integer} TODO
\item \code{sample.id} a \code{list} of 3 \code{list}. Each of the 3 
\code{list} contains a \code{list} of 12 entries each containing a string
of \code{character}, the name of the sample. TODO
}
}
\description{
TODO
}
\examples{

## Loading dataset
data(dataSimExample)

## TODO
}
\seealso{
\itemize{
\item \code{\link{runSim}} {for running a
simulation analysis using methylKit info as input}
}
}
\keyword{datasets}

