% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{fixSeed}
\alias{fixSeed}
\title{Fix seed value.}
\usage{
fixSeed(vSeed)
}
\arguments{
\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used.}
}
\value{
a \code{double}, the seed value.
}
\description{
Fix seed value when specified value is -1.
}
\examples{

## Return vSeed value when value is not -1
methInheritSim:::fixSeed(vSeed = 10)

## Return new value when value is -1
methInheritSim:::fixSeed(vSeed = -1)

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
