% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getStateData}
\alias{getStateData}
\title{TODO}
\usage{
getStateData(methInfo, m, block)
}
\arguments{
\item{methInfo}{is methylKitList from CTRL data.}

\item{m}{blocks}

\item{block}{a \code{integer} consecutive position from methInfo}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

