% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getSyntheticChr}
\alias{getSyntheticChr}
\title{Create a synthetic chromosome with the CTRL genome}
\usage{
getSyntheticChr(methInfo, m, nbCpG)
}
\arguments{
\item{methInfo}{is methylKitList of the CTRL data.}

\item{m}{\code{integer}, the number of blocks used for sampling.}

\item{nbCpG}{a \code{integer}, the number of consecutive CpG positions used
for sampling from \code{methInfo}.}
}
\value{
a \code{GRanges} object, the senthetic chromosome TODO
}
\description{
Create a synthetic chromosome with the sampling of a specified 
number of blocks and a specified number of consecutive CpG.
}
\examples{

## Load methyl information
data(samplesForChrSynthetic)

## Ensure results are reproducible
set.seed(32)
methylInheritanceSim:::getSyntheticChr(methInfo=samplesForChrSynthetic, 
m = 10, nbCpG = 20)

}
\author{
Pascal Belleau
}
\keyword{internal}

