% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methInheritSim.R
\docType{package}
\name{methInheritSim-package}
\alias{methInheritSim-package}
\alias{methInheritSim}
\title{methInheritSim: Simulating Whole-Genome Inherited Bisulphite Sequencing Data}
\description{
This package generates simulations of multigeneration of bisulfite data 
using a real control dataset.
}
\seealso{
\itemize{
\item \code{\link{runSim}} { for simulating a multigeneration methylation 
experiment with inheritance }
}
}
\author{
Pascal Belleau,
Astrid Deschênes and 
Arnaud Droit

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
