% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSim.R
\docType{package}
\name{methylInheritanceSim-package}
\alias{methylInheritanceSim-package}
\alias{methylInheritanceSim}
\title{methylInheritanceSim: Simulation TODO}
\description{
This package does a simulation of multigeneration of bisulfite data
}
\seealso{
\itemize{
\item \code{\link{runSim}} { TODO }
}
}
\author{
Pascal Belleau,
Astrid Deschênes and 
Arnaud Droit

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
