% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylPed.R
\docType{package}
\name{methylPed-package}
\alias{methylPed}
\alias{methylPed-package}
\title{methylPed: Simulation}
\description{
This package does a permutation analysis, based on Monte Carlo sampling,
for testing the hypothesis that the number of conserved differentially
methylated elements (sites, tiles or regions), between
several generations, is associated to an effect inherited from a treatment
and that stochastic effect can be dismissed.
}
\author{
Astrid Deschênes,
Pascal Belleau,
Mathieu Dalvai,
Janice Bailey and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{runPed}} { todo }
}
}
\keyword{package}

