% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimMethods.R
\name{reRunSim}
\alias{reRunSim}
\title{TODO}
\usage{
reRunSim(pathOut, fileGen, nbSynCHR, methData, nbBlock, lBlock, nbC,
  nbGeneration, vpDiff, vpDiffsd, vDiff, vInheritance, propInherite, rateDiff,
  minRate, propHetero, n, nbCores, keepDiff = FALSE)
}
\arguments{
\item{pathOut}{}

\item{fileGen}{}

\item{nbSynCHR}{}

\item{methData}{}

\item{nbBlock}{}

\item{lBlock}{}

\item{nbC}{}

\item{nbGeneration}{}

\item{vpDiff}{}

\item{vpDiffsd}{}

\item{vDiff}{}

\item{vInheritance}{}

\item{propInherite}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propHetero}{}

\item{nbCores}{}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}

