% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{restartSim}
\alias{restartSim}
\title{TODO}
\usage{
restartSim(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id, generation,
  res, rateDiff, minRate, propInherite, diffValue, propDiff, propDiffsd,
  propInheritance, propHetero, diffRes = NULL)
}
\arguments{
\item{pathOut}{}

\item{pref}{}

\item{k}{}

\item{nbCtrl}{}

\item{nbCase}{}

\item{treatment}{}

\item{sample.id}{}

\item{generation}{}

\item{res}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propInherite}{}

\item{diffValue}{}

\item{propDiff}{}

\item{propDiffsd}{}

\item{propInheritance}{}

\item{propHetero}{}

\item{diffRes}{(default \code{NULL})}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

