% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimMethods.R
\name{runSim}
\alias{runSim}
\title{TODO}
\usage{
runSim(pathOut, fileGen, nbSynCHR, methData, nbBlock, lBlock, nbC, nbGeneration,
  vpDiff, vpDiffsd, vDiff, vInheritance, propInherite, rateDiff, minRate,
  propHetero, n, nbCores, keepDiff = FALSE, vSeed = -1)
}
\arguments{
\item{pathOut}{}

\item{fileGen}{}

\item{nbSynCHR}{}

\item{methData}{}

\item{nbBlock}{}

\item{lBlock}{}

\item{nbC}{}

\item{nbGeneration}{}

\item{vpDiff}{}

\item{vpDiffsd}{}

\item{vDiff}{}

\item{vInheritance}{}

\item{propInherite}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propHetero}{}

\item{nbCores}{}

\item{keepDiff}{}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}

