% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{simInheritance}
\alias{simInheritance}
\title{TODO}
\usage{
simInheritance(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id,
  generation, stateInfo, rateDiff, minRate, propInherite, diffValue, propDiff,
  propDiffsd, propInheritance, propHetero, minReads = 10,
  maxPercReads = 99.9, context = "CpG", assembly = "Rnor_5.0",
  meanCov = 80, diffRes = NULL, saveGRanges = TRUE,
  saveMethylKit = TRUE, anaMethylKit = TRUE)
}
\arguments{
\item{pathOut}{}

\item{pref}{}

\item{k}{}

\item{nbCtrl}{}

\item{nbCase}{}

\item{treatment}{}

\item{sample.id}{}

\item{generation}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propInherite}{}

\item{diffValue}{}

\item{propDiff}{}

\item{propDiffsd}{}

\item{propInheritance}{}

\item{propHetero}{}

\item{diffRes}{(default \code{NULL})}

\item{res}{}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

