% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylPedInternalMethods.R
\name{estBetaAlpha}
\alias{estBetaAlpha}
\title{TODO}
\usage{
estBetaAlpha(valCtrl, minVal = 1e-06)
}
\arguments{
\item{valCtrl}{array with first value mean
and second}

\item{minVal}{float (default 1e-06)}
}
\value{
The alpha parameter of a Beta distribution
}
\description{
Estime the paramater alpha from the mean and the variance
of a beta distribution
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

