% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{estBetaBeta}
\alias{estBetaBeta}
\title{TODO}
\usage{
estBetaBeta(valCtrl, minVal = 1e-06)
}
\arguments{
\item{valCtrl}{array with first value mean
and second}

\item{minVal}{float (default 1e-06)}
}
\value{
The beta parameter of a Beta distribution
}
\description{
Estime the paramater beta from the mean and the variance
of a beta distribution
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

