% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getDiffCase}
\alias{getDiffCase}
\title{get a proportion c/t for a case at a differentially}
\usage{
getDiffCase(x, nb, sDiff, diffCase)
}
\arguments{
\item{x}{TODO}

\item{nb}{TODO}

\item{sDiff}{TODO}

\item{diffCase}{TODO}

\item{propDiffsd}{-c * exp(b * log(distance between two CpG))}

\item{b}{constant for distance}

\item{endLength}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO
data(dataSimExample)

getDiffCase(x = unlist(mcols(stateInfo)[1,3:4]), 
sDiff = 0.8, diffCase = round(6 * 0.9))

}
\author{
Pascal Belleau
}
\keyword{internal}

