% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getDiffCase}
\alias{getDiffCase}
\title{Get a proportion C/T for a case at a differentially}
\usage{
getDiffCase(x, nb, sDiff, diffCase)
}
\arguments{
\item{x}{a \code{vector} of \code{double} containing 3 entries: 
\itemize{
\item TODO
\item ToDO
\item ToDO 
}}

\item{nb}{TODO}

\item{sDiff}{TODO}

\item{diffCase}{TODO}

\item{propDiffsd}{TODO
-c * exp(b * log(distance between two CpG))}

\item{b}{constant for distance}

\item{endLength}{TODO}
}
\value{
a \code{vector} containing 3 entries:
\itemize{
\item TODO
\item TODO
\item TODO
}
}
\description{
TODO
}
\examples{

## Create vector containing 3 entries:
## 1 - TODO
## 2 - TODO
## 3 - TODO
x <- c(0.9814562, 0.0003607153, 0)

## Get the proportion of TODO
methylInheritanceSim:::getDiffCase(x=x, nb=6, sDiff = 0.8, 
diffCase = round(6 * 0.9))

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

