% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getDiffCase}
\alias{getDiffCase}
\title{TODO}
\usage{
getDiffCase(x, nb, sDiff, diffCase, propDiffsd)
}
\arguments{
\item{x}{}

\item{nb}{}

\item{sDiff}{}

\item{diffCase}{}

\item{propDiffsd}{-c * exp(b * log(distance between two CpG))}

\item{b}{constant for distance}

\item{endLength}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

