% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getDiffMeth}
\alias{getDiffMeth}
\title{TODO}
\usage{
getDiffMeth(stateInfo, rateDiff, minRate, propInherite, c = 1, b = -0.1,
  endLength = 1000)
}
\arguments{
\item{stateInfo}{TODO}

\item{rateDiff}{}

\item{minRate}{}

\item{propInherite}{}

\item{c}{a \code{double}, TODO Default: \code{1.0}.}

\item{b}{a \code{double}, TODO . Default: \code{-1e-01}.}

\item{endLength}{TODO . Default: \code{1000}.}
}
\value{
a \code{list} containing the 2 following elements:
\itemize{
\item \code{stateDiff} a \code{vector} of \code{integer} (\code{0} 
and \code{1}) with length corresponding the length of \code{stateInfo}.
The \code{vector}
indicates, using \code{1}, the positions where the CpG sites are
differentially methylated.
\item \code{stateInherite} a \code{vector} of \code{integer} (\code{0} and 
\code{1})
with length corresponding the length of \code{stateInfo}. The 
\code{vector}
indicates, using \code{1}, the positions where the CpG values are
inherited.
}
}
\description{
Identify the pos where the case are Diff meth and which one are heritable
}
\examples{

## TODO

## Load dataset containing TODO
data(dataSimExample)

diffRes <- methylInheritanceSim:::getDiffMeth(stateInfo = 
dataSimExample$stateInfo, rateDiff = 0.3, minRate = 0.3,
propInherite = 0.3)

}
\author{
Pascal Belleau
}
\keyword{internal}

