% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylPedInternalMethods.R
\name{getDiffMeth}
\alias{getDiffMeth}
\title{TODO}
\usage{
getDiffMeth(stateInfo, rateDiff, minRate, propInherite, c = 0.2, b = -0.2,
  endLength = 1000)
}
\arguments{
\item{stateInfo}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propInherite}{}

\item{c}{}

\item{b}{}

\item{endLength}{}
}
\value{
TODO
}
\description{
Identify the pos where the case are Diff meth and which one are heritable
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

