% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getDiffMeth}
\alias{getDiffMeth}
\title{TODO}
\usage{
getDiffMeth(stateInfo, rateDiff, minRate, propInherite, c = 1, b = -0.1,
  endLength = 1000)
}
\arguments{
\item{stateInfo}{}

\item{rateDiff}{}

\item{minRate}{}

\item{propInherite}{}

\item{b}{TODO . Default: \code{-1e-01}.}

\item{endLength}{TODO . Default: \code{1000}.}

\item{c.}{Default: \code{-1e-01}.}
}
\value{
TODO
}
\description{
Identify the pos where the case are Diff meth and which one are heritable
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

