% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getSim}
\alias{getSim}
\title{TODO}
\usage{
getSim(nbCtrl, nbCase, generation, stateInfo, stateDiff, diffValue, propDiff,
  propDiffsd = 0.1, propInheritance, propHetero)
}
\arguments{
\item{nbCtrl}{}

\item{nbCase}{}

\item{generation}{}

\item{stateInfo}{}

\item{stateDiff}{}

\item{diffValue}{}

\item{propDiff}{}

\item{propInheritance}{}

\item{propHetero}{}

\item{propDiffsd=0.1}{}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

