% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getSim}
\alias{getSim}
\title{TODO}
\usage{
getSim(nbCtrl, nbCase, generation, stateInfo, stateDiff, diffValue, propDiff,
  propDiffsd = 0.1, propInheritance, propHetero)
}
\arguments{
\item{nbCtrl}{a non-negative \code{integer}, the number of controls}

\item{nbCase}{a non-negative \code{integer}, the number of cases}

\item{generation}{TODO}

\item{stateInfo}{TODO}

\item{stateDiff}{}

\item{diffValue}{}

\item{propDiff}{}

\item{propDiffsd}{Default: \code{0.1}}

\item{propInheritance}{}

\item{propHetero}{}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

