% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{getSyntheticChr}
\alias{getSyntheticChr}
\title{Create a synthetic chromosome with the CTRL genome}
\usage{
getSyntheticChr(methInfo, m, nbCpG)
}
\arguments{
\item{methInfo}{is methylKitList of the CTRL data.}

\item{m}{\code{integer} the number of blocks}

\item{nbCpG}{a \code{integer} consecutive CpG position from methInfo}
}
\value{
TODO
}
\description{
Create a synthetic chromosome with the sampling of m blocks 
with nbCpG consecutive CpG
}
\examples{

## Load methyl information
data(samplesForChrSynthetic)
set.seed(32)
methylInheritanceSim:::getSyntheticChr(methInfo=samplesForChrSynthetic, m = 10, nbCpG = 20)


## TODO


}
\author{
Pascal Belleau
}
\keyword{internal}

