% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSim.R
\docType{package}
\name{methylInheritanceSim-package}
\alias{methylInheritanceSim-package}
\alias{methylInheritanceSim}
\title{methylInheritanceSim: Simulation of a multigeneration methylation case 
versus control experiment with inheritance relation using a real control 
dataset}
\description{
This package generates simulations of multigeneration of bisulfite data.
}
\seealso{
\itemize{
\item \code{\link{runSim}} { for simulating a multigeneration methylation 
experiment with inheritance }
}
}
\author{
Pascal Belleau,
Astrid Deschênes and 
Arnaud Droit

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
