% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSim.R
\docType{package}
\name{methylInheritanceSim-package}
\alias{methylInheritanceSim-package}
\alias{methylPed}
\title{methylInheritanceSim: Simulation}
\description{
This package does a simulation of multigeneration of bisulfite data
}
\author{
Astrid Deschênes,
Pascal Belleau,
Mathieu Dalvai,
Janice Bailey and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{runSim}} { todo }
    \item \code{\link{reRunSim}} { todo }
}
}
\keyword{package}

