% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimMethods.R
\name{runSim}
\alias{runSim}
\title{TODO}
\usage{
runSim(pathOut, fileGen, nbSynCHR, methData, nbBlock, lBlock, nbC, nbGeneration,
  vpDiff, vpDiffsd, vDiff, vInheritance, propInherite, rateDiff, minRate,
  propHetero, minReads = 10, maxPercReads = 99.9, context = "CpG",
  assembly = "Rnor_5.0", meanCov = 80, n, keepDiff = FALSE,
  saveGRanges = TRUE, saveMethylKit = TRUE, anaMethylKit = TRUE,
  nbCores = 1, vSeed = -1)
}
\arguments{
\item{pathOut}{the path where the the out put file is seved}

\item{fileGen}{a \code{string} include each output file. Each output 
file are 
composed with a type name (methylGR, methylObj, ...), _, fileGen (ex F1),
parameters of the simulation and ".rds".}

\item{nbSynCHR}{a \code{integer}, the number of distinct synthetics 
chromosomes. generate.}

\item{methData}{a \code{methylBase} object from \code{methylKit}, a real 
dataset used to sample block for the synthetic chromosome}

\item{nbBlock}{a \code{integer}, the number of blocks sample from methData 
genome to create the synthetic chromosome.}

\item{lBlock}{a \code{integer}, the number of the CpG in each block sampled}

\item{nbGeneration}{a \code{integer}, the number of generation simulate}

\item{vpDiff}{a positive \code{double} inferior ot \code{1}, the mean 
proportion of case simulate as differentially methylated when the site 
is selected as differentially methylated (this is like the penetrance).}

\item{vpDiffsd}{a \code{double} the variance of the parameter \code{vpDIff}}

\item{vDiff}{a positive \code{double} inferior ot \code{1}, in the case of 
a differentially methylated site a case 
differentially methylated the proportion of C/T follow a beta distribution 
where the mean is shifted of vDiff from the CTRL distribution}

\item{vInheritance}{a positive \code{double} inferior ot \code{1}, the 
proportion of case inherite the inherited sites.}

\item{propInherite}{a positive \code{double} inferior ot \code{1}, 
proportion of differentially methylated site
are inherated}

\item{rateDiff}{a positive \code{double} inferior ot \code{1}, the mean of 
the chance that a site is differentially 
methylated}

\item{minRate}{a positive \code{double} inferior ot \code{1}, the minimum 
number of diferentially methylated site}

\item{propHetero}{a positive \code{double} inferior ot \code{1}, the 
reduction vDiff for the intergeneration}

\item{minReads}{a positive \code{integer} Bases and regions having lower
coverage than this count are discarded. The parameter
correspond to the \code{lo.count} parameter in the  \code{methylKit} package.}

\item{maxPercReads}{a \code{double} between [0,100], the percentile of read
counts that is going to be used as upper cutoff. Bases ore regions
having higher
coverage than this percentile are discarded. Parameter used for both CpG
sites and tiles analysis. The parameter
correspond to the \code{hi.perc} parameter in the  \code{methylKit} package.}

\item{context}{Default: \code{"CpG"}}

\item{assembly}{Default: \code{"Rnor_5.0"}}

\item{meanCov}{Default: \code{80}}

\item{keepDiff}{Default: \code{FALSE}}

\item{saveGRanges}{Default: \code{TRUE}}

\item{saveMethylKit}{Default: \code{TRUE}}

\item{anaMethylKit}{Default: \code{TRUE}}

\item{nbCores}{a positive \code{integer}, the number of cores to use when
processing the analysis. Default: \code{1}.}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}

\item{vNbSample}{a \code{vector} of the number of CTRL and case in the the 
simulation dataset. In 
the simulation dataset the number of CTRL equal the number of Case. 
The number of CTRL do not need to be equal to the number of Case in
the real dataset}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO


}
\author{
Pascal Belleau
}

