% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimMethods.R
\name{runSim}
\alias{runSim}
\title{TODO}
\usage{
runSim(pathOut, fileGen, nbSynCHR, methData, nbBlock, lBlock, vNbSample,
  nbGeneration, vpDiff, vpDiffsd, vDiff, vInheritance, propInherite, rateDiff,
  minRate, propHetero, minReads = 10, maxPercReads = 99.9,
  context = "CpG", assembly = "Rnor_5.0", meanCov = 80, n,
  keepDiff = FALSE, saveGRanges = TRUE, saveMethylKit = TRUE,
  anaMethylKit = TRUE, nbCores = 1, vSeed = -1)
}
\arguments{
\item{pathOut}{the path where the the output file is seved}

\item{fileGen}{a \code{string} include each output file. Each output 
file are 
composed with a type name (methylGR, methylObj, ...), _, fileGen (ex F1),
parameters of the simulation and ".rds".}

\item{nbSynCHR}{a positive \code{integer}, the number of distinct synthetics 
chromosomes. generate.}

\item{methData}{is object of class \code{methylBase}, the CpG information
from controls (CTRL) that will be used to create the sythetic chromosome. 
The object can also contain information from cases but only the controls will
be used.}

\item{nbBlock}{\code{integer}, the number of blocks used for sampling.}

\item{lBlock}{a \code{integer}, the number of consecutive CpG positions used
for sampling from \code{methInfo}.}

\item{vNbSample}{a \code{vector} of positive \code{integer}, the number of 
CTRL and case in the the simulation dataset. In 
the simulation dataset the number of CTRL equal the number of Case. 
The number of CTRL do not need to be equal to the number of Case in
the real dataset}

\item{nbGeneration}{a positive \code{integer}, the number of generations}

\item{vpDiff}{a \code{double} superior to \code{0} and inferior or equal 
to \code{1}, the mean value for the proportion of samples that will have,
for a specific position, differentially methylated values. It can be 
interpreted as the penetrance.}

\item{vpDiffsd}{a non-negative \code{double}, the standard deviation 
associated to the \code{propDiff}.}

\item{vDiff}{a positive \code{double} between [0,1], the proportion of 
C/T for a case differentially methylated follow a beta distribution 
where the mean is shifted of \code{vDiff} from the CTRL distribution}

\item{vInheritance}{a positive \code{double} between [0,1], the 
proportion of cases that inherited differentially sites.}

\item{propInherite}{a positive \code{double} inferior to \code{1}, 
proportion of differentially methylated site
are inherated}

\item{rateDiff}{a positive \code{double} inferior to \code{1}, the mean of 
the chance that a site is differentially 
methylated}

\item{minRate}{a positive \code{double} inferior to \code{1}, the minimum 
rate of differentially methylated sites}

\item{propHetero}{a positive \code{double} between [0,1], the 
reduction of vDiff for the second and following generation}

\item{minReads}{a positive \code{integer} Bases and regions having lower
coverage than this count are discarded. The parameter
correspond to the \code{lo.count} parameter in the \code{methylKit} package.}

\item{maxPercReads}{a \code{double} between [0,100], the percentile of read
counts that is going to be used as upper cutoff. Bases ore regions
having higher
coverage than this percentile are discarded. Parameter used for both CpG
sites and tiles analysis. The parameter
correspond to the \code{hi.perc} parameter in the  \code{methylKit} package.}

\item{context}{methylation context string, ex: CpG,CpH,CHH, etc. (default:CpG)}

\item{assembly}{string that determines the genome assembly. Ex: mm9,hg18 etc.
Default: \code{"Rnor_5.0"}}

\item{meanCov}{a positive \code{integer} represent the mean of the coverage
at the CpG site Default: \code{80}}

\item{n}{a positive \code{integer} the number of simulation for each 
parameters (\code{vNbSample}, \code{vpDiff}, \code{vDiff} and
\code{vInheritance})}

\item{keepDiff}{\code{logical} if true, the differentially methyled sites
will be the same for each parameters ( \code{vpDiff}, 
\code{vDiff} and \code{vInheritance}). Default: \code{FALSE}}

\item{saveGRanges}{\code{logical} if true, save a list of \code{nbGeneration}
GRangesList. Each GRangeList contain \code{vNbSample} GRanges which contain
for each CpG site: the position, the coverage and the proportion of the C/T.
A files treatment_... which contains the position of the CTRL and Case are 
save too. Default: \code{TRUE}}

\item{saveMethylKit}{Default: \code{TRUE}}

\item{anaMethylKit}{Default: \code{TRUE}}

\item{nbCores}{a positive \code{integer}, the number of cores to use when
processing the analysis. Default: \code{1}.}

\item{vSeed}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO pathOut where to write file ?
## Load methyl information
data(samplesForChrSynthetic)

\dontrun{runSim(pathOut = "testData", fileGen = "F1", nbSynCHR = 1, 
methData = samplesForChrSynthetic, nbBlock = 10, lBlock = 20,
vNbSample = c(3), nbGeneration = 3, vpDiff = c(0.9), 
vpDiffsd = c(0.1), vDiff = c(0.8), 
vInheritance = c(0.5), propInherite = 0.3,
rateDiff = 0.01, minRate = 0.01, propHetero = 0.5, n = 5, 
nbCores= 1, vSeed = 32)}

}
\author{
Pascal Belleau
}

