% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSim.R
\docType{data}
\name{samplesForChrSynthetic}
\alias{samplesForChrSynthetic}
\title{All samples information, formated by \code{methylKit}, in a
\code{methylBase} format (for demo purpose).}
\format{A \code{methylBase} object contains the information for one generation.
Each sample information is
stored in a \code{methylRaw} object. There is \code{methylRaw} objects
 (6 controls and 6 cases).}
\usage{
data(samplesForChrSynthetic)
}
\value{
A \code{methylBase} contains the information for one generation.
Each sample information is stored in a \code{methylRaw} object. 
There is \code{methylRaw} objects
(6 controls and 6 cases).
}
\description{
The object is a \code{methylBase}.
There is 12 samples (6 controls and 6 cases). Each
sample information is stored in a \code{methylRaw} object.
}
\details{
This dataset can be
used to test the \code{runSim} function.
}
\examples{

## Loading dataset
data(samplesForChrSynthetic)

## Run a permutation analysis
\dontrun{runSim(TODO)}

}
\seealso{
\itemize{
\item \code{\link{runSim}} {for running a
simulation analysis using methylKit info entry}
}
}
\keyword{datasets}

