% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{simInheritance}
\alias{simInheritance}
\title{TODO}
\usage{
simInheritance(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id,
  generation, stateInfo, propDiff, propDiffsd, diffValue, propInheritance,
  rateDiff, minRate, propInherite, propHetero, minReads = 10,
  maxPercReads = 99.9, context = "CpG", assembly = "Rnor_5.0",
  meanCov = 80, diffRes = NULL, saveGRanges = TRUE,
  saveMethylKit = TRUE, runAnalysis = TRUE)
}
\arguments{
\item{pathOut}{TODO}

\item{pref}{TODO}

\item{k}{TODO}

\item{nbCtrl}{a positive \code{integer}, the number of controls.}

\item{nbCase}{a positive \code{integer}, the number of cases.}

\item{treatment}{TODO}

\item{sample.id}{TODO}

\item{generation}{a positive \code{integer}, the number of generations.}

\item{stateInfo}{TODO}

\item{propDiff}{TODO}

\item{propDiffsd}{TODO}

\item{diffValue}{TODO}

\item{propInheritance}{TODO}

\item{rateDiff}{TODO}

\item{minRate}{TODO}

\item{propInherite}{TODO}

\item{propHetero}{TODO}

\item{minReads}{TODO}

\item{maxPercReads}{TODO}

\item{context}{TODO. Default: \code{"CpG"}.}

\item{assembly}{TODO. Default: \code{"Rnor_5.0"}.}

\item{meanCov}{TODO. Default: \code{80}.}

\item{diffRes}{TODO. Default: \code{NULL}.}

\item{saveGRanges}{TODO. Default: \code{TRUE}.}

\item{saveMethylKit}{TODO. Default: \code{TRUE}.}

\item{runAnalysis}{TODO. Default: \code{TRUE}.}
}
\value{
\code{0} indicating that the function has been successful.
}
\description{
TODO
}
\examples{

temp_dir <- "test_simInheritance"
data(dataSimExample)

\dontrun{methylInheritanceSim:::simInheritance(pathOut = temp_dir,
pref = paste0("S1_", "6_0.9_0.8_0.5"),
k = 1, nbCtrl = 6, nbCase = 6, 
treatment = dataSimExample$treatment, 
sample.id = dataSimExample$sample.id,
generation = 3, 
stateInfo = dataSimExample$stateInfo,
propDiff = 0.9, propDiffsd = 0.1,
diffValue = 0.8,
propInheritance = 0.5,
rateDiff = 0.3, minRate = 0.3,
propInherite = 0.3, 
propHetero = 0.5,
runAnalysis = FALSE
)}

## Delete temp_dir
\dontrun{if (dir.exists(temp_dir)) {
unlink(temp_dir, recursive = TRUE, force = FALSE)
}}

}
\author{
Pascal Belleau
}
\keyword{internal}

