% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{simInheritance}
\alias{simInheritance}
\title{TODO}
\usage{
simInheritance(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id,
  generation, stateInfo, rateDiff, minRate, propInherite, diffValue, propDiff,
  propDiffsd, propInheritance, propHetero, minReads = 10,
  maxPercReads = 99.9, context = "CpG", assembly = "Rnor_5.0",
  meanCov = 80, diffRes = NULL, saveGRanges = TRUE,
  saveMethylKit = TRUE, anaMethylKit = TRUE)
}
\arguments{
\item{pathOut}{TODO}

\item{pref}{TODO}

\item{k}{TODO}

\item{nbCtrl}{a positive \code{integer}, the number of controls.}

\item{nbCase}{a positive \code{integer}, the number of cases.}

\item{treatment}{a \code{vector} containing 0 and 1 denoting 
which samples are controls and which samples are cases. TODO}

\item{sample.id}{TODO}

\item{generation}{a positive \code{integer}, the number of generations.}

\item{stateInfo}{TODO}

\item{rateDiff}{a positive \code{double} inferior to \code{1}, the mean of 
the probability that a site is differentially methylated.}

\item{minRate}{a positive \code{double} inferior to \code{1}, the minimum 
rate of differentially methylated sites. Ensure that the total number of
differentially methylated sites is superior to \code{minRate}.}

\item{propInherite}{a positive \code{double} between [0,1], the 
proportion of cases that inherited differentially methylated sites.}

\item{diffValue}{TODO}

\item{propDiff}{TODO}

\item{propDiffsd}{TODO}

\item{propInheritance}{TODO}

\item{propHetero}{TODO}

\item{minReads}{a  positive \code{integer}, bases and regions having lower
coverage inferior to \code{minReads} are discarded. This parameter
correspond to the \code{lo.count} parameter in the \code{methylKit} package.}

\item{maxPercReads}{TODO}

\item{context}{a string of \code{character}, a short description of 
the methylation context. As exemples: Cpg, CpH, CHH, etc.. 
Default: \code{"CpG"}.}

\item{assembly}{a string of \code{character}, TODO. 
Default: \code{"Rnor_5.0"}.}

\item{meanCov}{TODO. Default: \code{80}.}

\item{diffRes}{TODO. Default: \code{NULL}.}

\item{saveGRanges}{a \code{logical}, TODO. Default: \code{TRUE}.}

\item{saveMethylKit}{a \code{logical}, TODO.. Default: \code{TRUE}.}

\item{anaMethylKit}{a \code{logical}, TODO. Default: \code{TRUE}.}
}
\value{
\code{0} indicating that the function has been successful.
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

