% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{simInheritance}
\alias{simInheritance}
\title{TODO}
\usage{
simInheritance(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id,
  generation, stateInfo, rateDiff, minRate, propInherite, diffValue, propDiff,
  propDiffsd, propInheritance, propHetero, minReads = 10,
  maxPercReads = 99.9, context = "CpG", assembly = "Rnor_5.0",
  meanCov = 80, diffRes = NULL, saveGRanges = TRUE,
  saveMethylKit = TRUE, anaMethylKit = TRUE)
}
\arguments{
\item{pathOut}{TODO}

\item{pref}{TODO}

\item{k}{TODO}

\item{nbCtrl}{a positive \code{integer}, the number of controls.}

\item{nbCase}{a positive \code{integer}, the number of cases.}

\item{treatment}{TODO}

\item{sample.id}{TODO}

\item{generation}{a positive \code{integer}, the number of generations.}

\item{stateInfo}{TODO}

\item{rateDiff}{}

\item{minRate}{TODO}

\item{propInherite}{TODO}

\item{diffValue}{TODO}

\item{propDiff}{TODO}

\item{propDiffsd}{TODO}

\item{propInheritance}{TODO}

\item{propHetero}{TODO}

\item{minReads}{TODO}

\item{maxPercReads}{TODO}

\item{context}{TODO. Default: \code{"CpG"}}

\item{assembly}{TODO}

\item{meanCov}{TODO. Default: \code{80}.}

\item{diffRes}{TODO. Default: \code{NULL}.}

\item{saveGRanges}{TODO. Default: \code{TRUE}.}

\item{saveMethylKit}{TODO. Default: \code{TRUE}.}

\item{anaMethylKit}{TODO. Default: \code{TRUE}.}
}
\value{
\code{0} indicating that the function has been successful.
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}

