% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylInheritanceSimInternalMethods.R
\name{simInheritance}
\alias{simInheritance}
\title{TODO}
\usage{
simInheritance(pathOut, pref, k, nbCtrl, nbCase, treatment, sample.id,
  generation, stateInfo, propDiff, propDiffsd, diffValue, propInheritance,
  rateDiff, minRate, propInherite, propHetero, minReads, maxPercReads, context,
  assembly, meanCov, diffRes, saveGRanges, saveMethylKit, runAnalysis)
}
\arguments{
\item{pathOut}{TODO}

\item{pref}{TODO}

\item{k}{TODO}

\item{nbCtrl}{a positive \code{integer}, the number of controls.}

\item{nbCase}{a positive \code{integer}, the number of cases.}

\item{treatment}{TODO}

\item{sample.id}{TODO}

\item{generation}{a positive \code{integer}, the number of generations
simulated.}

\item{stateInfo}{TODO}

\item{propDiff}{TODO}

\item{propDiffsd}{TODO}

\item{diffValue}{TODO}

\item{propInheritance}{TODO}

\item{rateDiff}{a positive \code{double} inferior to \code{1}, the mean of 
the chance that a site is differentially methylated.}

\item{minRate}{a non-negative \code{double} inferior to \code{1}, the 
minimum rate for differentially methylated sites.
Default: \code{0.01}.}

\item{propInherite}{a non-negative \code{double} inferior or equal 
to \code{1}, 
the proportion of differentially methylated regions that 
are inherated.}

\item{propHetero}{a non-negative \code{double} between [0,1], the 
reduction of \code{vDiff} for the second and following generations.}

\item{minReads}{a positive \code{integer}, sites and regions having lower
coverage than this count are discarded. The parameter
corresponds to the \code{lo.count} parameter in 
the \code{methylKit} package.}

\item{maxPercReads}{a \code{double} between [0,100], the percentile of read
counts that is going to be used as upper cutoff. Sites and regions
having higher
coverage than \code{maxPercReads} are discarded. This parameter is used for 
both CpG sites and tiles analysis. The parameter
correspond to the \code{hi.perc} parameter in the \code{methylKit} package.}

\item{context}{a string of \code{character}, the short description of the 
methylation context, such as "CpG", "CpH", "CHH", etc..}

\item{assembly}{a string of \code{character}, the short description of the 
genome assembly, such as "mm9", "hg18", etc..}

\item{meanCov}{a positive \code{integer}, the mean of the coverage
at the simulated CpG sites.}

\item{diffRes}{TODO when \code{NULL}}

\item{saveGRanges}{TODO.}

\item{saveMethylKit}{TODO.}

\item{runAnalysis}{TODO.}
}
\value{
\code{0} indicating that the function has been successful.
}
\description{
TODO
}
\examples{

temp_dir <- "test_simInheritance"
data(dataSimExample)

\dontrun{methylInheritanceSim:::simInheritance(pathOut = temp_dir,
pref = paste0("S1_", "6_0.9_0.8_0.5"),
k = 1, nbCtrl = 6, nbCase = 6, 
treatment = dataSimExample$treatment, 
sample.id = dataSimExample$sample.id,
generation = 3, 
stateInfo = dataSimExample$stateInfo,
propDiff = 0.9, propDiffsd = 0.1,
diffValue = 0.8,
propInheritance = 0.5,
rateDiff = 0.3, minRate = 0.3,
propInherite = 0.3, 
propHetero = 0.5,
saveGRanges = FALSE,
saveMethylKit = FALSE,
runAnalysis = FALSE
)}

## Delete temp_dir
\dontrun{if (dir.exists(temp_dir)) {
unlink(temp_dir, recursive = TRUE, force = FALSE)
}}

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

